+++ packages/frontend/src/components/common/PivotTable/ValueCell.tsx
@@ -14,6 +14,7 @@ import { FC, useCallback, useEffect, useMemo, useState } from 'react';
 import { getColorFromRange, readableColor } from '../../../utils/colorUtils';
 import { getConditionalRuleLabel } from '../Filters/configs';
 import Cell, { CellProps } from './Cell';
+import DashboardValueCellMenu from './DashboardValueCellMenu';
 import { usePivotTableCellStyles } from './tableStyles';
 import ValueCellMenu from './ValueCellMenu';
 
@@ -27,6 +28,8 @@ interface ValueCellProps extends CellProps {
         colIndex: number,
         rowIndex: number,
     ) => Record<string, ResultValue>;
+    isDashboard?: boolean;
+    tileUuid?: string;
 }
 
 const SMALL_TEXT_LENGTH = 30;
@@ -38,7 +41,8 @@ const ValueCell: FC<ValueCellProps> = ({
     colIndex,
     conditionalFormattings,
     getUnderlyingFieldValues,
-
+    tileUuid,
+    isDashboard,
     ...rest
 }) => {
     const conditionalFormatting = useMemo(() => {
@@ -99,38 +103,46 @@ const ValueCell: FC<ValueCellProps> = ({
         };
     }, [handleCopy, isMenuOpen]);
 
-    return (
-        <ValueCellMenu
-            rowIndex={rowIndex}
-            colIndex={colIndex}
-            opened={isMenuOpen}
-            item={item}
-            value={value}
-            getUnderlyingFieldValues={getUnderlyingFieldValues}
-            onCopy={handleCopy}
-            onOpen={() => setIsMenuOpen(true)}
-            onClose={() => setIsMenuOpen(false)}
+    const cell = (
+        <Cell
+            withValue={!!formattedValue}
+            withNumericValue={isNumericItem(item)}
+            className={cx(
+                {
+                    [classes.conditionalFormatting]: conditionalFormatting,
+                    [classes.withLargeText]:
+                        formattedValue &&
+                        formattedValue?.length > SMALL_TEXT_LENGTH,
+                },
+                rest.className,
+            )}
+            data-conditional-formatting={!!conditionalFormatting}
+            data-copied={clipboard.copied}
+            tooltipContent={conditionalFormatting?.tooltipContent}
+            {...rest}
         >
-            <Cell
-                withValue={!!formattedValue}
-                withNumericValue={isNumericItem(item)}
-                className={cx(
-                    {
-                        [classes.conditionalFormatting]: conditionalFormatting,
-                        [classes.withLargeText]:
-                            formattedValue &&
-                            formattedValue?.length > SMALL_TEXT_LENGTH,
-                    },
-                    rest.className,
-                )}
-                data-conditional-formatting={!!conditionalFormatting}
-                data-copied={clipboard.copied}
-                tooltipContent={conditionalFormatting?.tooltipContent}
-                {...rest}
-            >
-                {formattedValue}
-            </Cell>
-        </ValueCellMenu>
+            {formattedValue}
+        </Cell>
+    );
+
+    const commonCellMenuProps = {
+        rowIndex: rowIndex,
+        colIndex: colIndex,
+        opened: isMenuOpen,
+        item: item,
+        value: value,
+        getUnderlyingFieldValues: getUnderlyingFieldValues,
+        onCopy: handleCopy,
+        onOpen: () => setIsMenuOpen(true),
+        onClose: () => setIsMenuOpen(false),
+    };
+
+    return isDashboard && tileUuid ? (
+        <DashboardValueCellMenu tileUuid={tileUuid} {...commonCellMenuProps}>
+            {cell}
+        </DashboardValueCellMenu>
+    ) : (
+        <ValueCellMenu {...commonCellMenuProps}>{cell}</ValueCellMenu>
     );
 };