+++ packages/frontend/src/components/common/ResourceIcon/index.tsx
@@ -68,29 +68,23 @@ export const getChartIcon = (chartType: ChartKind | undefined) => {
     switch (chartType) {
         case undefined:
         case ChartKind.VERTICAL_BAR:
-            return <IconBox icon={IconChartBar} color="blue.8" />;
+            return IconChartBar;
         case ChartKind.HORIZONTAL_BAR:
-            return (
-                <IconBox
-                    icon={IconChartBar}
-                    color="blue.8"
-                    transform="rotate(90)"
-                />
-            );
+            return IconChartBar;
         case ChartKind.LINE:
-            return <IconBox icon={IconChartLine} color="blue.8" />;
+            return IconChartLine;
         case ChartKind.SCATTER:
-            return <IconBox icon={IconChartDots} color="blue.8" />;
+            return IconChartDots;
         case ChartKind.AREA:
-            return <IconBox icon={IconChartArea} color="blue.8" />;
+            return IconChartArea;
         case ChartKind.MIXED:
-            return <IconBox icon={IconChartAreaLine} color="blue.8" />;
+            return IconChartAreaLine;
         case ChartKind.PIE:
-            return <IconBox icon={IconChartPie} color="blue.8" />;
+            return IconChartPie;
         case ChartKind.TABLE:
-            return <IconBox icon={IconTable} color="blue.8" />;
+            return IconTable;
         case ChartKind.BIG_NUMBER:
-            return <IconBox icon={IconSquareNumber1} color="blue.8" />;
+            return IconSquareNumber1;
         default:
             return assertUnreachable(
                 chartType,
@@ -99,6 +93,18 @@ export const getChartIcon = (chartType: ChartKind | undefined) => {
     }
 };
 
+export const ChartIcon: FC<{ chartType: ChartKind | undefined }> = ({
+    chartType,
+}) => (
+    <IconBox
+        icon={getChartIcon(chartType)}
+        color="blue.8"
+        transform={
+            chartType === ChartKind.HORIZONTAL_BAR ? `"rotate(90)"` : undefined
+        }
+    />
+);
+
 export const ResourceIcon: FC<ResourceIconProps> = ({ item }) => {
     switch (item.type) {
         case ResourceViewItemType.DASHBOARD:
@@ -106,7 +112,7 @@ export const ResourceIcon: FC<ResourceIconProps> = ({ item }) => {
         case ResourceViewItemType.SPACE:
             return <IconBox icon={IconFolder} color="violet.8" />;
         case ResourceViewItemType.CHART:
-            return getChartIcon(item.data.chartType);
+            return <ChartIcon chartType={item.data.chartType} />;
         default:
             return assertUnreachable(item, 'Resource type not supported');
     }